edición de categorías

+++ botica/src/main/java/com/example/botica/repository/MarcaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.botica.model.Marca;
 
-public interface MarcaRepository extends JpaRepository<Marca, Long> {
+public interface MarcaRepository extends JpaRepository<Marca, Integer> {
 }
