+++ botica/src/main/java/com/example/botica/repository/PresentacionRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.botica.model.Presentacion;
 
-public interface PresentacionRepository extends JpaRepository<Presentacion, Long> {
+public interface PresentacionRepository extends JpaRepository<Presentacion, Integer> {
 }
