+++ botica/src/main/java/com/example/botica/repository/SintomaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.botica.model.Sintoma;
 
-public interface SintomaRepository extends JpaRepository<Sintoma, Long> {
+public interface SintomaRepository extends JpaRepository<Sintoma, Integer> {
 }
